fix: split batch JSON tests for pkg-maint limits
