+++ pkg/cli/submit/batch_test.go
@@ -780,181 +780,6 @@ func TestSubmitBatch_HumanSuccessOutputTruncatesLongWorkLists(t *testing.T) {
 	}
 }
 
-func TestSubmitBatch_JSONSuccessOutputIncludesRequiredFields(t *testing.T) {
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(factoryapi.UpsertWorkRequestResponse{
-			RequestId: "batch-json-1",
-			TraceId:   "trace-json-1",
-			Works: []factoryapi.UpsertWorkRequestSubmittedWork{
-				{Name: "alpha", WorkTypeName: "task", WorkId: "work-alpha"},
-				{Name: "beta", WorkTypeName: "review", WorkId: ""},
-			},
-		})
-	}))
-	defer srv.Close()
-
-	path := writeBatchFile(t, `{
-		"requestId": "batch-json-1",
-		"type": "FACTORY_REQUEST_BATCH",
-		"works": [
-			{"name": "alpha", "workTypeName": "task", "payload": {"secret": "hidden"}},
-			{"name": "beta", "workTypeName": "review", "payload": {"secret": "hidden"}}
-		],
-		"relations": [{"type": "DEPENDS_ON", "sourceWorkName": "alpha", "targetWorkName": "beta"}]
-	}`)
-
-	var out bytes.Buffer
-	err := SubmitBatch(BatchConfig{
-		Args:      []string{path},
-		Server:    mustServerBase(t, srv.URL),
-		SessionID: "session-json",
-		JSON:      true,
-		Output:    &out,
-	})
-	if err != nil {
-		t.Fatalf("SubmitBatch: %v", err)
-	}
-
-	var got BatchSubmitJSONResult
-	if err := json.Unmarshal(out.Bytes(), &got); err != nil {
-		t.Fatalf("decode JSON output: %v\n%s", err, out.String())
-	}
-	if got.RequestID != "batch-json-1" {
-		t.Fatalf("requestId = %q, want batch-json-1", got.RequestID)
-	}
-	if got.TraceID != "trace-json-1" {
-		t.Fatalf("traceId = %q, want trace-json-1", got.TraceID)
-	}
-	if got.WorkCount != 2 {
-		t.Fatalf("workCount = %d, want 2", got.WorkCount)
-	}
-	if got.RelationCount != 1 {
-		t.Fatalf("relationCount = %d, want 1", got.RelationCount)
-	}
-	if got.SessionID != "session-json" {
-		t.Fatalf("sessionId = %q, want session-json", got.SessionID)
-	}
-	if got.EndpointPath != "/factory-sessions/session-json/work-requests/batch-json-1" {
-		t.Fatalf("endpointPath = %q, want session-scoped work-requests path", got.EndpointPath)
-	}
-	if got.BatchSource != batchSourceFile {
-		t.Fatalf("batchSource = %q, want %q", got.BatchSource, batchSourceFile)
-	}
-	if len(got.Works) != 2 {
-		t.Fatalf("works len = %d, want 2", len(got.Works))
-	}
-	if got.Works[0].Name != "alpha" || got.Works[0].WorkTypeName != "task" || got.Works[0].WorkID != "work-alpha" {
-		t.Fatalf("works[0] = %#v, want alpha/task/work-alpha", got.Works[0])
-	}
-	if got.Works[1].Name != "beta" || got.Works[1].WorkTypeName != "review" || got.Works[1].WorkID != "" {
-		t.Fatalf("works[1] = %#v, want beta/review without workId", got.Works[1])
-	}
-	if got.DryRun {
-		t.Fatal("dryRun should be omitted or false on HTTP success")
-	}
-	if strings.Contains(out.String(), "hidden") {
-		t.Fatalf("JSON output leaked batch payload:\n%s", out.String())
-	}
-}
-
-func TestSubmitBatch_JSONDryRunIncludesSummaryWithoutTraceIDUnlessPresent(t *testing.T) {
-	path := writeBatchFile(t, validBatchJSON("batch-json-dry", "alpha"))
-
-	var out bytes.Buffer
-	err := SubmitBatch(BatchConfig{
-		Args:   []string{path},
-		DryRun: true,
-		JSON:   true,
-		Server: "http://127.0.0.1:1",
-		Output: &out,
-	})
-	if err != nil {
-		t.Fatalf("SubmitBatch: %v", err)
-	}
-
-	var got BatchSubmitJSONResult
-	if err := json.Unmarshal(out.Bytes(), &got); err != nil {
-		t.Fatalf("decode JSON output: %v\n%s", err, out.String())
-	}
-	if !got.DryRun {
-		t.Fatal("dryRun = false, want true")
-	}
-	if got.RequestID != "batch-json-dry" {
-		t.Fatalf("requestId = %q, want batch-json-dry", got.RequestID)
-	}
-	if got.TraceID != "" {
-		t.Fatalf("traceId = %q, want omitted when absent from input", got.TraceID)
-	}
-	if got.WorkCount != 1 {
-		t.Fatalf("workCount = %d, want 1", got.WorkCount)
-	}
-	if got.RelationCount != 0 {
-		t.Fatalf("relationCount = %d, want 0", got.RelationCount)
-	}
-	if got.SessionID != "~default" {
-		t.Fatalf("sessionId = %q, want ~default", got.SessionID)
-	}
-	if got.EndpointPath != "/factory-sessions/~default/work-requests/batch-json-dry" {
-		t.Fatalf("endpointPath = %q, want default session dry-run path", got.EndpointPath)
-	}
-	if got.BatchSource != batchSourceFile {
-		t.Fatalf("batchSource = %q, want %q", got.BatchSource, batchSourceFile)
-	}
-	if len(got.WorkNames) != 1 || got.WorkNames[0] != "alpha" {
-		t.Fatalf("workNames = %#v, want [alpha]", got.WorkNames)
-	}
-	if len(got.Works) != 0 {
-		t.Fatalf("works = %#v, want omitted on dry-run", got.Works)
-	}
-}
-
-func TestSubmitBatch_JSONDryRunIncludesTraceIDWhenPresentInInput(t *testing.T) {
-	path := writeBatchFile(t, `{
-		"requestId": "batch-json-trace",
-		"type": "FACTORY_REQUEST_BATCH",
-		"currentChainingTraceId": "trace-from-input",
-		"works": [{"name": "alpha", "workTypeName": "task", "payload": {"title": "A"}}]
-	}`)
-
-	var out bytes.Buffer
-	err := SubmitBatch(BatchConfig{
-		Args:   []string{path},
-		DryRun: true,
-		JSON:   true,
-		Server: "http://127.0.0.1:1",
-		Output: &out,
-	})
-	if err != nil {
-		t.Fatalf("SubmitBatch: %v", err)
-	}
-
-	var got BatchSubmitJSONResult
-	if err := json.Unmarshal(out.Bytes(), &got); err != nil {
-		t.Fatalf("decode JSON output: %v\n%s", err, out.String())
-	}
-	if got.TraceID != "trace-from-input" {
-		t.Fatalf("traceId = %q, want trace-from-input", got.TraceID)
-	}
-}
-
-func TestSubmitBatch_ValidationErrorDoesNotEmitSuccessJSON(t *testing.T) {
-	var out bytes.Buffer
-	err := SubmitBatch(BatchConfig{
-		Args:   []string{`{"requestId":"bad","type":"FACTORY_REQUEST_BATCH","works":[]}`},
-		JSON:   true,
-		Server: "http://127.0.0.1:1",
-		Output: &out,
-	})
-	if err == nil {
-		t.Fatal("expected validation error")
-	}
-	if out.Len() != 0 {
-		t.Fatalf("stdout = %q, want empty on validation failure", out.String())
-	}
-}
-
 func TestSubmitBatch_UsesDocsExampleStartupWorkFile(t *testing.T) {
 	path := testutil.MustRepoPath(t, "docs/examples/startup-work.json")
 	err := SubmitBatch(BatchConfig{
